% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_natural}
\alias{check_natural}
\title{The \code{checks.R} script is intended to include \code{check_...} functions
used within various session functions. They are not intended to be
exported (though can be seen with \code{MTM:::check_...}). They are
generally meant to be ignored by students, and used to provide
informative errors when the functions that \emph{are} exported are
misapplied.}
\usage{
check_natural(N)
}
\description{
Exported functions, meant to be used directly by students in
practicals, should have their arguments verified by one or
more \code{check_...} functions.
}
\details{
In general, \code{check_...}s should:
\itemize{
\item concern one feature at a time
\item invisibly return the checked value
\item be applied via pipe
}
}
