% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{coord_max}
\alias{coord_max}
\title{Cartesian Coordinates, 0 to Max}
\usage{
coord_max(
  xmax = NA,
  ymax = NA,
  xlim = c(0, xmax),
  ylim = c(0, ymax),
  expand = FALSE,
  ...
)
}
\arguments{
\item{xmax}{alternate parameterization for xlim}

\item{ymax}{alternate parameterization for ylim}

\item{xlim, ylim}{Limits for the x and y axes.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian}}
  \describe{
    \item{\code{xlim,ylim}}{Limits for the x and y axes.}
    \item{\code{default}}{Is this the default coordinate system? If \code{FALSE} (the default),
then replacing this coordinate system with another one creates a message alerting
the user that the coordinate system is being replaced. If \code{TRUE}, that warning
is suppressed.}
    \item{\code{clip}}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. In most cases, the default of \code{"on"} should not be changed,
as setting \code{clip = "off"} can cause unexpected results. It allows
drawing of data points anywhere on the plot, including in the plot margins. If
limits are set via \code{xlim} and \code{ylim} and some data points fall outside those
limits, then those data points may show up in places such as the axes, the
legend, the plot title, or the plot margins.}
  }}
}
\description{
Cartesian Coordinates, 0 to Max
}
