% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{ggplot.igraph}
\alias{ggplot.igraph}
\title{GGplot for \link{igraph}}
\usage{
\method{ggplot}{igraph}(data, mapping = aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{an \link{igraph} object}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggplot]{ggplot2::ggplot}}
  \describe{
    \item{\code{mapping}}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}
    \item{\code{environment}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Used prior to tidy
evaluation.}
  }}
}
\description{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} specialization for \link{igraph} objects
}
\details{
Applies \code{\link[=as.data.table.igraph]{as.data.table.igraph()}} to \code{data} and then passes it into
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
