% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{launch}
\alias{launch}
\title{launches one of the Modern Techniques in Modelling (MTM) short course
sessions}
\usage{
launch(session)
}
\arguments{
\item{session, }{a string or integer, indicating which session to launch}
}
\description{
launches one of the Modern Techniques in Modelling (MTM) short course
sessions
}
