% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_internal.R
\name{match.call.defaults}
\alias{match.call.defaults}
\title{\link[base:match.call]{base::match.call} Extension}
\usage{
match.call.defaults(
  definition = sys.function(sys.parent()),
  call = sys.call(sys.parent()),
  expand.dots = TRUE,
  envir = parent.frame(2L)
)
}
\arguments{
\item{definition}{a function, by default the function from which
    \code{match.call} is called.  See details.}

\item{call}{an unevaluated call to the function specified by
    \code{definition}, as generated by \code{\link[base]{call}}.}

\item{expand.dots}{logical.  Should arguments matching \code{\dots}
    in the call be included or left as a \code{\dots} argument?}

\item{envir}{an environment, from which the \code{\dots} in \code{call}
    are retrieved, if any.}
}
\description{
Produces a \code{match.call} with defaults
also included.
}
