% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{network_active_edges}
\alias{network_active_edges}
\title{Extract Active Edges}
\usage{
network_active_edges(dt)
}
\arguments{
\item{dt}{a \link{data.table} resulting from \code{\link[=as.data.table.igraph]{as.data.table.igraph()}},
applied to an \link{igraph} with both vertex and edge attributes \code{state}.}
}
\description{
Extract Active Edges
}
