% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{network_animate}
\alias{network_animate}
\title{create an animated plot of transmission on the population}
\usage{
network_animate(networks)
}
\arguments{
\item{networks}{a list of igraphs}
}
\value{
gganimate object
}
\description{
create an animated plot of transmission on the population
}
\seealso{
gganimate::anim_save
}
