% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_data.R
\docType{data}
\name{network_data}
\alias{network_data}
\alias{network_warmup_vaccine_ordered}
\alias{network_warmup_vaccine_random}
\alias{network_unstructured_set}
\alias{network_structured_set}
\title{Networks Session Data}
\format{
An object of class \code{igraph} of length 100.

An object of class \code{igraph} of length 100.

An object of class \code{list} of length 300.

An object of class \code{list} of length 300.
}
\usage{
network_warmup_vaccine_ordered

network_warmup_vaccine_random

network_unstructured_set

network_structured_set
}
\description{
Data used in the \emph{Networks} session scripts and tutorials.
}
\keyword{MTM}
\keyword{data}
\keyword{datasets}
