% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network_dReedFrost}
\alias{network_dReedFrost}
\title{Compute 1-step transition for Reed-Frost Network Model}
\usage{
network_dReedFrost(t, y, parms, ...)
}
\arguments{
\item{t, }{the time step; ignored in the Reed-Frost model}

\item{y}{an \link{igraph} network, the initial state of the population}

\item{parms}{parameters passed to \code{func}.}

\item{...}{other arguments passed to \code{func}.}
}
\value{
an \code{igraph} representing the transitions that occur and edges involved
}
\description{
Compute 1-step transition for Reed-Frost Network Model
}
\examples{
require(MTM)
parms <- list(N=30, p=0.1)
set.seed(8675309)
sirpop <- network_build(parms)
onestep <- network_dReedFrost(0, sirpop, parms)

# the initial state
network_quickplot(sirpop)

# the changes
network_quickplot(onestep)

}
\seealso{
network_solve
}
