% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{network_edge_data}
\alias{network_edge_data}
\title{Get Edge Data}
\usage{
network_edge_data(dt)
}
\arguments{
\item{dt}{\link{data.table} object from \code{\link[=as.data.table.igraph]{as.data.table.igraph()}}}
}
\description{
Extract edge data from \link{data.table}s made from \link{igraph}s
}
\details{
Intended for use with \code{\link[=geom_edge]{geom_edge()}}. Extracts the unique edges,
their ends, and any attribute data.
}
