% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network_flatten}
\alias{network_flatten}
\title{Aggregate Network into States}
\usage{
network_flatten(y, one = is.igraph(y))
}
\arguments{
\item{y}{an \code{igraph}, representing a population OR a list of such \code{igraph}s}
}
\description{
Aggregate Network into States
}
