% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network_percolate}
\alias{network_percolate}
\title{Transform a network by percolation}
\usage{
network_percolate(parms, network = network_build(parms))
}
\arguments{
\item{parms}{a \code{list(N=integer, p=probability)}; the Reed-Frost model parameters.}

\item{network, }{an igraph object, such as produced by \link{network_build};
for when you want to repeatedly percolate the same base network with
different probabilities.}
}
\value{
a new igraph object, with potentially fewer edges. it's 'E()$draw' attribute
will be == 0 (i.e., always transmit)
}
\description{
Transform a network by percolation
}
\examples{
require(MTM)
parms <- list(N=30, p=0.2)

# repeat these next steps a few times, w/ different `p` to see
# to see how `network_percolate()` works.
sirpop <- network_percolate(within(parms, p <- 0.05))
network_quickplot(sirpop, edgeargs = list(), vertexargs = list())

}
