% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{network_plot_histograms}
\alias{network_plot_histograms}
\title{plot summary of a series of simulations}
\usage{
network_plot_histograms(s.dt)
}
\arguments{
\item{s.dt}{a data.table, of the same structure as returned by \code{\link{network_sample_reed_frost}}}
}
\value{
a patchwork'd ggplot object
}
\description{
plot summary of a series of simulations
}
