% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{network_plot_series}
\alias{network_plot_series}
\title{plot the time series samples}
\usage{
network_plot_series(s_dt, layer_order = c("S", "R", "I"))
}
\arguments{
\item{s_dt}{a \code{data.table}, of the same structure as returned by
\code{\link{network_sample_reed_frost}}}
}
\description{
plot the time series samples
}
