% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{network_quickplot}
\alias{network_quickplot}
\title{Demo Network Plotter}
\usage{
network_quickplot(
  ig,
  values,
  labels,
  edgeargs = if (simple) list(color = "grey90") else list(),
  vertexargs = if (simple) list(color = "black") else list(),
  simple = FALSE
)
}
\arguments{
\item{ig}{an \link{igraph}}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}

\item{edgeargs}{a list of arguments to \code{\link[=geom_edge]{geom_edge()}}}

\item{vertexargs}{a list of arguments to \code{\link[=geom_vertex]{geom_vertex()}}}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Provides a quick plotter with most elements constrained
}
