% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network_summarize}
\alias{network_summarize}
\title{Summarize Network Runs}
\usage{
network_summarize(dt)
}
\arguments{
\item{dt, }{an object coercable by \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.
Expected to have columns \code{sample},\code{t}, and \code{R}.
See \code{\link[=network_sample]{network_sample()}} return value}
}
\description{
Calculates final size and duration of
simulated epidemics on networks.
}
