% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network_update}
\alias{network_update}
\title{compute a new network from initial states & changes}
\usage{
network_update(y, dy)
}
\arguments{
\item{y, }{an \code{igraph}, the initial network state}

\item{dy, }{an \code{igraph}, the vertex states that change & edges that were
activated leading to those changes}
}
\value{
a new igraph, representing the state changes applied
}
\description{
compute a new network from initial states & changes
}
\details{
Note: this function is only exported so that students
can see the internals easily. It is intended to be used
internally to other functions in the package, where the
arguments are already assured, so it does not include
\code{check_...} steps, which would be too expensive to repeatedly
calculate.
}
\examples{
population <- network_build_basic(N = 30, p = 0.1)
# TODO

}
