% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{network_vertex_data}
\alias{network_vertex_data}
\title{Get Vertex Data}
\usage{
network_vertex_data(dt)
}
\arguments{
\item{dt}{\link{data.table} object from \code{\link[=as.data.table.igraph]{as.data.table.igraph()}}}
}
\description{
Extract vertex data from \link{data.table}s made from \link{igraph}s
}
\details{
Intended for use with \code{\link[=geom_vertex]{geom_vertex()}}. Extracts the unique vertices,
their positions, and any attribute data.
}
