% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{patchwork_grid}
\alias{patchwork_grid}
\title{Patchwork Grid}
\usage{
patchwork_grid(plist)
}
\arguments{
\item{plist}{a list of named lists of \link{ggplot2} objects, representing a set
of plots to layout as a grid.}
}
\description{
Provides a quick way to convert a named list structure
(possibly nested) of plots into a tagged grid arrangement of those plots.
}
