% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_2D_density}
\alias{plot_2D_density}
\title{Multi-dimensional density plot}
\usage{
plot_2D_density(
  dt,
  margins = setdiff(colnames(s.dt), "sample")[1:2],
  scales = mget(sprintf("scale_\%s_\%s", c("x", "y"), margins))
)
}
\arguments{
\item{dt}{a \link{data.table}, with a \code{sample} column and at least other two columns
of properties to visualise}
}
\value{
a \link{patchwork}'d \link{ggplot2} object
}
\description{
Multi-dimensional density plot
}
