% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{question_freeresponse}
\alias{question_freeresponse}
\title{a \code{learnr::question_text} wrapper}
\usage{
question_freeresponse(text, reply, ...)
}
\arguments{
\item{text, }{see \code{learnr::question_text}}

\item{reply, }{a string, allowing \emph{bold} and \emph{italic} markdown indicators,
which provides an example answer}
}
\description{
a \code{learnr::question_text} wrapper
}
\details{
\code{learnr::quiz} framework doesn't support free-response style
Q&A. This function provides a way for participants to provide an open-end
response, which isn't explicitly checked, and to receive an example
reply after submitting their answer
}
