% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_internal.R
\name{rejig}
\alias{rejig}
\title{Generic Function Wrapper}
\usage{
rejig(FUN, ..., .ENV = environment(FUN))
}
\arguments{
\item{FUN}{the function to wrap}

\item{...}{the new defaults}

\item{.ENV}{the environment for the resulting
copy-function (i.e. where any variables will be
evaluated). NB, the default (\code{environment(FUN)}) is
mostly convenient, but can be dangerous e.g. by
replacing an important function}
}
\value{
the new function
}
\description{
provides a convenience function for
producing duplicate functions with different
defaults
}
