% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reminder}
\alias{reminder}
\title{Reminders for Scripts}
\usage{
reminder(
  custom = NULL,
  text =
    "For the script below, our general guidance is to run the code,\\ncheck the results in the console or plots tab, and then\\nanswer the associated questions yourself or amongst a small group.\\n\\nAnswering some of the questions will entail changing function\\narguments or filling in missing arguments. These are marked\\nwith `TODO(...)` placeholders.\\n\\nFor some of those, we suggest relevant functions with\\n'@hint: ?FUNCTION', as in `?FUNCTION` will prompt you with the\\ndocumentation of a function that may be useful\\n\\nThere are some items marked with @aside; skip these until you\\nhave completed all other elements of a particular activity."
)
}
\arguments{
\item{custom, }{text to display after generic reminder}

\item{text, }{the generic reminder}
}
\description{
Reminders for Scripts
}
\examples{
reminder("Some exhortation.")
}
