% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sampling_maxprevalence}
\alias{sampling_maxprevalence}
\title{Find Max Prevalence in SIR ODE}
\usage{
sampling_maxprevalence(
  parameters,
  dModel = sampling_dSIR_betagamma,
  plot_results = FALSE
)
}
\arguments{
\item{parameters}{a list of the parameters appropriate to use with \code{dModel}}

\item{dModel}{either \code{\link[=sampling_dSIR_betagamma]{sampling_dSIR_betagamma()}} or \code{\link[=sampling_dSIR_gammaR0]{sampling_dSIR_gammaR0()}}}

\item{plot_results}{plot (to the active device) the "S" and "I" compartments}
}
\value{
the maximum "I" value for the solved ODE, a numeric scalar
}
\description{
Given appropriate parameter values, solves an SIR model,
then finds the maximum "I" value. Uses fixed population
of c(S=99, I=1, R=0), and runs out to \code{t=50}.
}
\examples{
require(MTM)
print(sampling_maxprevalence(ist(beta = 0.4, gamma = 0.2)))

}
\seealso{
Other sampling: 
\code{\link{sampling_MC_plot}()},
\code{\link{sampling_dSIR_betagamma}()}
}
\concept{sampling}
