% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_plotting.R
\name{sampling_MC_plot}
\alias{sampling_MC_plot}
\title{Plot Monte Carlo Results}
\usage{
sampling_MC_plot(data)
}
\arguments{
\item{data}{a \code{data.frame}, with columns \code{R0}, \code{gamma}, and \code{max.prev}}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Produces a plot of Monte Carlo Simulation Results for SIR model
}
\seealso{
Other sampling: 
\code{\link{sampling_dSIR_betagamma}()},
\code{\link{sampling_maxprevalence}()}
}
\concept{sampling}
