% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{SIRcolors}
\alias{SIRcolors}
\title{defines reference colors used in plots throughout MTM course}
\format{
An object of class \code{character} of length 3.
}
\usage{
SIRcolors
}
\description{
defines reference colors used in plots throughout MTM course
}
\keyword{datasets}
