% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochdisc_functions.R
\name{stochdisc_dReedFrost}
\alias{stochdisc_dReedFrost}
\title{Compute 1-step transition for Reed-Frost Chain Binomial Model}
\usage{
stochdisc_dReedFrost(t, y, parms, ...)
}
\arguments{
\item{t, }{the time step; ignored in the Reed-Frost model}

\item{y}{a vector of initial conditions}

\item{parms}{a list of parameters to pass to \code{func}.}

\item{...}{ignored}
}
\value{
an list; first element, a vector of state changes,
optional second element, a vector of other global parameter values
}
\description{
Compute 1-step transition for Reed-Frost Chain Binomial Model
}
\examples{
require(MTM)
ps <- list(N=30, p=0.1)
sirpop <- c(S = ps$N - 1, I = 1, R = 0)
dpop <- stochdisc_dReedFrost(y = sirpop, parms = ps)
dpop
# ... random results each time
stochdisc_dReedFrost(y = sirpop, parms = ps)

}
\seealso{
stochdisc_solve

Other stochdisc: 
\code{\link{stochdisc_sample}()},
\code{\link{stochdisc_solve}()}
}
\concept{stochdisc}
