% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochdisc_functions.R
\name{stochdisc_solve}
\alias{stochdisc_solve}
\title{Solve Stochastic, Discrete Time Models}
\usage{
stochdisc_solve(
  y = c(S = parms$N - 1, I = 1, R = 0),
  times = NULL,
  func = stochdisc_dReedFrost,
  parms,
  ...
)
}
\arguments{
\item{y, }{a vector of initial conditions}

\item{times, }{a numeric vector, the times to report the state; cast
to \code{0:as.integer(max(times))}. If \code{NULL} (the default) runs until extinct.}

\item{func, }{a function that computes the change in state; see
\code{\link[deSolve:ode]{deSolve::ode()}} for details of the function signature.}

\item{parms, }{a list of parameters to pass to \code{func}.}

\item{..., }{ignored}
}
\value{
a \code{\link[data.table:data.table]{data.table::data.table()}} with columns \code{t} and others
corresponding to the state variables in \code{y}.
}
\description{
Given initial conditions, step function,
and parameters, compute the time series of the model.
}
\examples{
require(MTM)
ps <- list(N=100, p = 0.02)
series.dt <- stochdisc_solve(parms = ps)
series.dt
# ... random results every time
series.dt <- stochdisc_solve(parms = ps)
series.dt

}
\seealso{
Other stochdisc: 
\code{\link{stochdisc_dReedFrost}()},
\code{\link{stochdisc_sample}()}
}
\concept{stochdisc}
