% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plotting.R
\name{theme_network}
\alias{theme_network}
\title{network plotting theme}
\usage{
theme_network(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
a \link[ggplot2:theme]{ggplot2::theme}, with defaults for network plots
}
