% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TODO}
\alias{TODO}
\title{Provides a placeholder that students can fill in}
\usage{
TODO(message)
}
\arguments{
\item{message, }{a string; explains what students should do at some point in
scripts / tutorials}
}
\description{
Provides a placeholder that students can fill in
}
