% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\docType{data}
\name{tutorial.list}
\alias{tutorial.list}
\title{The available tutorials}
\format{
An object of class \code{character} of length 1.
}
\usage{
tutorial.list
}
\description{
The available tutorials
}
\keyword{datasets}
