% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MTM}
\alias{MTM}
\title{MTM}
\usage{
MTM(Y, Xf = NULL, K = NULL, resCov = list(type = "UN", df0 = 0, S0 =
  diag(0, ncol(as.matrix(Y)))), nIter = 110, burnIn = 10, thin = 2,
  saveAt = "", tolD = 1e-05)
}
\arguments{
\item{Y}{Phenotypes}

\item{Xf}{Incidence matrix}

\item{K}{Kernel}

\item{resCov}{Residual variance}

\item{nIter}{Number of iterations}

\item{burnIn}{}

\item{thin}{}

\item{saveAt}{}

\item{tolD}{}
}
\value{
Model
}
\description{
\code{MTM} does inference of parameters from the posterior distribution of
a multivariate normal model with structured co-variance matrices.
}

